requirements.txt
# -e git+https://github.com/isu-enterprise/isu.enterprise#egg=isu.enterprise
psycopg2
pyramid
chameleon
pyramid_chameleon
pyramid_zcml
pyramid_debugtoolbar
cornice
icc.mvw
